Fix #11350: MegaMenu rtl and start/end facets
