+++ docs/14_0_0/components/tieredmenu.md
@@ -94,7 +94,7 @@ TieredMenu supports `start` and `end` facets to control UI output or add logos o
     </f:facet>
     <p:menuitem label="Logout" action="#{bean.logout}"/>
     <f:facet name="end">
-       <p:button value="Logout" icon="pi pi-fw pi-sign-out" styleClass="ui-button-info"/>
+       <p:button value="Quit" icon="pi pi-fw pi-sign-out" styleClass="ui-button-info"/>
     </f:facet>
 </p:tieredMenu>
 ```
